# packages/backend-sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/events": {
      "get": {
        "operationId": "ProjectCoreController_streamEvents",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/users/me/projects": {
      "get": {
        "operationId": "ProjectCoreController_findUserProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSerialized"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/users/me/projects/search": {
      "get": {
        "operationId": "ProjectCoreController_findUserProjectsForSearch",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSearchResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects": {
      "post": {
        "operationId": "ProjectCoreController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}": {
      "get": {
        "operationId": "ProjectCoreController_findById",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectCoreController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectCoreController_delete",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}/suggested-users": {
      "get": {
        "operationId": "ProjectCoreController_getSuggestedUsers",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPartialSerialized"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}/history": {
      "get": {
        "operationId": "ProjectCoreController_findHistoryById",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSecretsVersionSerialized"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}/analytics/secrets-pushed": {
      "post": {
        "operationId": "ProjectCoreController_trackSecretsPushed",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}/encrypted-secrets-keys": {
      "post": {
        "operationId": "ProjectCoreController_addEncryptedSecretsKey",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEncryptedSecretsKeyBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{projectId}/members/{memberId}": {
      "delete": {
        "operationId": "ProjectCoreController_removeMember",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectCoreController_updateMemberRole",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cryptly",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "project-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "project-api-key"
      }
    },
    "schemas": {
...
      "ProjectMemberSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "publicKey": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "admin"
            ]
          }
        },
        "required": [
          "id",
          "avatarUrl",
          "displayName",
          "role"
        ]
      },
      "ProjectSettingsSerialized": {
        "type": "object",
        "properties": {
          "revealOn": {
            "type": "string",
            "enum": [
              "always",
              "hover",
              "never"
            ]
          }
        },
        "required": [
          "revealOn"
        ]
      },
      "ProjectSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMemberSerialized"
            }
          },
          "encryptedSecretsKeys": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "encryptedSecrets": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettingsSerialized"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "members",
          "encryptedSecretsKeys",
          "encryptedSecrets",
          "settings",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProjectSearchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "encryptedSecretsKeys": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "encryptedSecrets": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "encryptedSecretsKeys",
          "encryptedSecrets"
        ]
      },
      "ProjectSettingsBody": {
        "type": "object",
        "properties": {
          "revealOn": {
            "type": "string",
            "enum": [
              "always",
              "hover",
              "never"
            ]
          }
        },
        "required": [
          "revealOn"
        ]
      },
      "CreateProjectBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "encryptedSecrets": {
            "type": "string"
          },
          "encryptedSecretsKeys": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettingsBody"
          }
        },
        "required": [
          "name",
          "encryptedSecrets",
          "encryptedSecretsKeys",
          "settings"
        ]
      },
      "UserPartialSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "publicKey": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "avatarUrl",
          "displayName"
        ]
      },
      "ProjectSecretsVersionSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/UserPartialSerialized"
          },
          "encryptedSecrets": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "author",
          "encryptedSecrets",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProjectBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "encryptedSecrets": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettingsBody"
          }
        }
      },
      "AddEncryptedSecretsKeyBody": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "encryptedSecretsKey": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "encryptedSecretsKey"
        ]
      },
      "UpdateMemberRoleBody": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "admin"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}